% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{check_suggested_packages}
\alias{check_suggested_packages}
\title{Check for required suggested packages}
\usage{
check_suggested_packages()
}
\description{
Check for required suggested packages
}
\keyword{internal}
