% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{coef.wendy}
\alias{coef.wendy}
\title{Extract coefficients from a wendy object}
\usage{
\method{coef}{wendy}(object, ...)
}
\arguments{
\item{object}{A wendy object}

\item{...}{Additional arguments (ignored)}
}
\value{
Numeric vector of estimated parameters
}
\description{
Extract coefficients from a wendy object
}
