% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsindy.R
\name{coef.wsindy}
\alias{coef.wsindy}
\title{Extract the coefficient matrix from a wsindy object}
\usage{
\method{coef}{wsindy}(object, ...)
}
\arguments{
\item{object}{A wsindy object}

\item{...}{Additional arguments (ignored)}
}
\value{
Term-by-state numeric coefficient matrix \code{W}
}
\description{
Extract the coefficient matrix from a wsindy object
}
