% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\docType{data}
\name{default_control}
\alias{default_control}
\title{Default control for solveWendy}
\format{
An object of class \code{list} of length 24.
}
\usage{
default_control
}
\description{
Default control for solveWendy
}
\keyword{datasets}
