% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{fitted.wendy}
\alias{fitted.wendy}
\title{Fitted values from a wendy object}
\usage{
\method{fitted}{wendy}(object, ...)
}
\arguments{
\item{object}{A wendy object}

\item{...}{Additional arguments (ignored)}
}
\value{
Numeric vector of fitted values
}
\description{
Fitted values from a wendy object
}
