% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp_predict}
\alias{gp_predict}
\title{Predict function values from a fitted GP}
\usage{
gp_predict(fit, tt_star)
}
\arguments{
\item{fit}{A gp_fit object from \code{gp_fit_1d}.}

\item{tt_star}{Prediction time points.}
}
\value{
List with mean (vector) and var (vector of marginal variances).
}
\description{
Predict function values from a fitted GP
}
