% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{matern52_Kdd}
\alias{matern52_Kdd}
\title{Matern 5/2 derivative kernel cov(f'(tt1), f'(tt2))}
\usage{
matern52_Kdd(tt1, tt2, sigma2, ell)
}
\arguments{
\item{tt1, tt2}{Numeric vectors of time points.}

\item{sigma2}{Signal variance (> 0).}

\item{ell}{Length scale (> 0).}
}
\value{
Matrix of size length(tt1) x length(tt2).
}
\description{
Matern 5/2 derivative kernel cov(f'(tt1), f'(tt2))
}
\keyword{internal}
