% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{matern52_Kfd}
\alias{matern52_Kfd}
\title{Matern 5/2 cross-covariance cov(f(tt1), f'(tt2))}
\usage{
matern52_Kfd(tt1, tt2, sigma2, ell)
}
\arguments{
\item{tt1, tt2}{Numeric vectors of time points.}

\item{sigma2}{Signal variance (> 0).}

\item{ell}{Length scale (> 0).}
}
\value{
Matrix of size length(tt1) x length(tt2).
}
\description{
Derivative is with respect to the second argument.
}
\keyword{internal}
