% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{matern52_matrix}
\alias{matern52_matrix}
\title{Matern 5/2 covariance matrix}
\usage{
matern52_matrix(tt1, tt2, sigma2, ell)
}
\arguments{
\item{tt1, tt2}{Numeric vectors of time points.}

\item{sigma2}{Signal variance (> 0).}

\item{ell}{Length scale (> 0).}
}
\value{
Matrix of size length(tt1) x length(tt2).
}
\description{
Matern 5/2 covariance matrix
}
