% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{nugget}
\alias{nugget}
\title{Observation noise covariance (nugget)}
\usage{
nugget(n, sigma2_n)
}
\arguments{
\item{n}{Number of observations.}

\item{sigma2_n}{Noise variance.}
}
\value{
Diagonal matrix n x n.
}
\description{
Observation noise covariance (nugget)
}
