% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{print.wendy}
\alias{print.wendy}
\title{Print method for wendy objects}
\usage{
\method{print}{wendy}(x, ...)
}
\arguments{
\item{x}{A wendy object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for wendy objects
}
