% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsindy.R
\name{print.wsindy}
\alias{print.wsindy}
\title{Print method for wsindy objects}
\usage{
\method{print}{wsindy}(x, digits = 4, ...)
}
\arguments{
\item{x}{A wsindy object returned by [solveWSINDy()]}

\item{digits}{Number of significant digits for the displayed coefficients}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for wsindy objects
}
