% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{rel_err}
\alias{rel_err}
\title{Calculate the relative error of two vectors}
\usage{
rel_err(x, y)
}
\arguments{
\item{x}{Vector}

\item{y}{Vector}
}
\value{
Scaler
}
\description{
Calculate the relative error of two vectors
}
