% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{residuals.wendy}
\alias{residuals.wendy}
\title{Extract residuals from a wendy object}
\usage{
\method{residuals}{wendy}(object, ...)
}
\arguments{
\item{object}{A wendy object}

\item{...}{Additional arguments (ignored)}
}
\value{
Numeric vector of weak residuals
}
\description{
Extract residuals from a wendy object
}
