% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{residuals_weighted}
\alias{residuals_weighted}
\title{Extract corrected residuals from a wendy object}
\usage{
residuals_weighted(object, ...)
}
\arguments{
\item{object}{A wendy object}

\item{...}{Additional arguments (ignored)}
}
\value{
Numeric vector of weak residuals
}
\description{
Extract corrected residuals from a wendy object
}
