% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsindy.R
\name{summary.wsindy}
\alias{summary.wsindy}
\title{Summary method for wsindy objects}
\usage{
\method{summary}{wsindy}(object, ...)
}
\arguments{
\item{object}{A wsindy object returned by [solveWSINDy()]}

\item{...}{Additional arguments (ignored)}
}
\value{
An object of class \code{summary.wsindy}
}
\description{
Summary method for wsindy objects
}
