% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wendy_filter.R
\name{estimate_u0}
\alias{estimate_u0}
\title{Estimate the initial condition using Euler-MacLaurin Trapezoid correction on the function g(t) := φ(t)f(p,u,t) + φ′(t)u(t)}
\usage{
estimate_u0(U, f_, J_u, J_t, tt, p, test_function_params)
}
\description{
Estimate the initial condition using Euler-MacLaurin Trapezoid correction on the function g(t) := φ(t)f(p,u,t) + φ′(t)u(t)
}
