% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wendy_filter.R
\name{wendy_erts}
\alias{wendy_erts}
\title{Estimate the state using the RTS smoother}
\usage{
wendy_erts(
  U,
  f_,
  J_u,
  J_t,
  tt,
  p,
  test_function_params,
  dF_dt_ = NULL,
  d2F_dt2_ = NULL,
  d3F_dt3_ = NULL,
  sigma = NULL
)
}
\description{
Estimate the state using the RTS smoother
}
